+++ src/Controllers/GlobalData/CurrencyController.php
@@ -2,10 +2,14 @@
 
 namespace JtlWooCommerceConnector\Controllers\GlobalData;
 
+use Exception;
 use Jtl\Connector\Core\Model\Currency as CurrencyModel;
 use Jtl\Connector\Core\Model\Identity;
+use JtlWooCommerceConnector\Controllers\AbstractBaseController;
+use JtlWooCommerceConnector\Integrations\Plugins\Wpml\Wpml;
+use JtlWooCommerceConnector\Integrations\Plugins\Wpml\WpmlCurrency;
 
-class CurrencyController
+class CurrencyController extends AbstractBaseController
 {
     public const ISO                = 'woocommerce_currency';
     public const SIGN_POSITION      = 'woocommerce_currency_pos';
@@ -13,14 +17,23 @@ class CurrencyController
     public const THOUSAND_DELIMITER = 'woocommerce_price_thousand_sep';
 
     /**
-     * @return CurrencyModel
+     * @return CurrencyModel[]
+     * @throws Exception
      */
-    public function pull(): CurrencyModel
+    public function pull(): array
     {
-        $iso = \get_woocommerce_currency();
+        $currencies = [];
 
-        return
-            (new CurrencyModel())
+        $wpmlCurrency = $this->getPluginsManager()
+            ->get(Wpml::class)
+            ->getComponent(WpmlCurrency::class);
+
+        if ($wpmlCurrency->canUseMultiCurrency()) {
+            $currencies = $wpmlCurrency->getCurrencies();
+        } else {
+            $iso = \get_woocommerce_currency();
+
+            $currencies[] = (new CurrencyModel())
                 ->setId(new Identity(\strtolower($iso)))
                 ->setName($iso)
                 ->setDelimiterCent(\get_option(self::THOUSAND_DELIMITER, ''))
@@ -29,11 +42,15 @@ class CurrencyController
                 ->setNameHtml(\get_woocommerce_currency_symbol())
                 ->setHasCurrencySignBeforeValue(\get_option(self::SIGN_POSITION, '') === 'left')
                 ->setIsDefault(true);
+        }
+
+        return $currencies;
     }
 
     /**
      * @param array $currencies
      * @return array
+     * @throws Exception
      */
     public function push(array $currencies): array
     {
@@ -53,6 +70,12 @@ class CurrencyController
             );
 
             break;
+        }
+
+        $wpml = $this->getPluginsManager()->get(Wpml::class);
+
+        if ($wpml->canBeUsed()) {
+            $wpml->getComponent(WpmlCurrency::class)->setCurrencies(...$currencies);
         }
 
         return $currencies;