+++ src/Controllers/Product/ProductVariationController.php
@@ -356,11 +356,15 @@ class ProductVariationController extends AbstractBaseController
         foreach ($pushedVariations as $variation) {
             foreach ($variation->getValues() as $variationValue) {
                 foreach ($variation->getI18ns() as $variationI18n) {
-                    if (!$this->util->isWooCommerceLanguage($variationI18n->getLanguageISO())) {
+                    if ($this->skipNotDefaultLanguage($variationI18n->getLanguageISO())) {
                         continue;
                     }
 
                     foreach ($variationValue->getI18ns() as $i18n) {
+                        if ($this->skipNotDefaultLanguage($i18n->getLanguageISO())) {
+                            continue;
+                        }
+
                         $metaKey                =
                             'attribute_pa_' . \wc_sanitize_taxonomy_name(
                                 \substr(
@@ -395,6 +399,26 @@ class ProductVariationController extends AbstractBaseController
 
         return $updatedAttributeKeys;
     }
+
+    /**
+     * @throws InvalidArgumentException
+     * @throws \Exception
+     */
+    protected function skipNotDefaultLanguage(string $wawiLanguageIso): bool
+    {
+        if ($this->wpml->canBeUsed()) {
+            if ($this->wpml->convertLanguageToWawi($this->wpml->getDefaultLanguage()) !== $wawiLanguageIso) {
+                return true;
+            }
+        } else {
+            if (!$this->util->isWooCommerceLanguage($wawiLanguageIso)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+
     // </editor-fold>
 
     // <editor-fold defaultstate="collapsed" desc="Methods">